Add an export npm script
Quite hacky at this point, though works.

Approach: uses puppeteer to open a URL and press Space, while taking
screenshots. We assume that at each transition the URL changes. Once it
stops changing, we assume there are no more slides.

Plenty of problems:
- requires `node`;
- assumes that the URL is visible without authentication and from the
environment we are running the script from (there can be proxies, etc.);
- no error handling – not sure what will happen if something breaks;
- the assumption that the URL will change quickly enough is weak, I can
see how it either finish too early or lead to an infinite loop. Knowing
the number of slides in advance will be better;
- assumes we have less than 1000 slides, so that we can take care of
padding in file names. Seems the safest assumption compared to the ones
above :) Can be lifted if we know the number of slides in advance;
- Resolution is hard-coded at Full HD – 1920:1080. Quite typical
projector resolution these days.